requirements.txt
# requirements.txt

# Core HTTP and networking
httpx>=0.25.0
requests>=2.31.0
dnspython>=2.4.0
scapy>=2.5.0

# CLI and output
rich>=13.0.0
click>=8.1.0

# Proxy and session analysis
mitmproxy>=10.0.0

# GUI dependencies (optional)
PySide6>=6.5.0

# Web preview (optional)
fastapi>=0.100.0
uvicorn>=0.23.0
slowapi>=0.1.7

# Packaging
pyinstaller>=6.0.0

# Async support
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0
